+++ Framework/include/untrusted_util.h
@@ -7,7 +7,7 @@
 typedef void* secure_connection;
 
 namespace untrusted_util {
-    long time_elapsed_ms(struct timeval start, struct timeval end);
+    double time_elapsed_ms(struct timeval start, struct timeval end);
     void debug_printbuf(uint8_t* buf, size_t len);
 
     int init_server(const int server_port);
