+++ Framework/src/LibTrusted/trusted_util.cpp
@@ -2,12 +2,12 @@
 
 #include "sgx_tprotected_fs.h"
 
-long trusted_util::time_elapsed_ms(untrusted_time start, untrusted_time end) {
+double trusted_util::time_elapsed_ms(untrusted_time start, untrusted_time end) {
     long secs_used, micros_used;
 
     secs_used = (end.tv_sec - start.tv_sec); //avoid overflow by subtracting first
     micros_used = ((secs_used*1000000) + end.tv_usec) - (start.tv_usec);
-    return micros_used / 1000;
+    return (double)micros_used / 1000;
 }
 
 // SGX_FILE*
