+++ Framework/src/LibUntrusted/untrusted_util.cpp
@@ -118,10 +118,10 @@ void untrusted_util::debug_printbuf(uint8_t* buf, size_t len) {
     printf("\n");
 }
 
-long untrusted_util::time_elapsed_ms(struct timeval start, struct timeval end) {
+double untrusted_util::time_elapsed_ms(struct timeval start, struct timeval end) {
     long secs_used, micros_used;
 
     secs_used = (end.tv_sec - start.tv_sec); //avoid overflow by subtracting first
     micros_used = ((secs_used*1000000) + end.tv_usec) - (start.tv_usec);
-    return micros_used / 1000;
+    return (double)micros_used / 1000.0;
 }
